% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\title{Run all build tasks: Sass, Rust, and TypeScript}
\usage{
build(package_name = NULL, path = ".")
}
\arguments{
\item{package_name}{Name of the package. If NULL, attempts to detect from DESCRIPTION.}

\item{path}{Path to the package root. Default: "."}
}
\description{
Run all build tasks: Sass, Rust, and TypeScript
}
