% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_rs}
\alias{build_rs}
\title{Build Rust WASM package}
\usage{
build_rs(package_name = get_pkg_name(path), dir = "srcrs", path = ".")
}
\arguments{
\item{package_name}{Name of the package (used to name output directory)}

\item{dir}{Directory containing Rust sources. Default: "srcrs"}

\item{path}{Path to the package root. Default: "."}
}
\description{
Build Rust WASM package
}
