% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_sass}
\alias{build_sass}
\title{Build Sass assets for the package}
\usage{
build_sass(package_name = get_pkg_name(path), dir = "srcsass", path = ".")
}
\arguments{
\item{package_name}{Name of the package (used to locate inst and out directories)}

\item{dir}{Directory containing Sass sources. Default: "srcsass"}

\item{path}{Path to the package root. Default: "."}
}
\description{
Build Sass assets for the package
}
