% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_ts}
\alias{build_ts}
\title{Build TypeScript assets (depends on Rust WASM)}
\usage{
build_ts(package_name = get_pkg_name(path), dir = "srcts", path = ".")
}
\arguments{
\item{package_name}{Name of the package (used to upgrade WASM dependency)}

\item{dir}{Directory containing TypeScript sources. Default: "srcts"}

\item{path}{Path to the package root. Default: "."}
}
\description{
Build TypeScript assets (depends on Rust WASM)
}
