% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_input_init.R, R/generic_output_init.R
\name{generic_component_init}
\alias{generic_component_init}
\alias{new_input_component}
\alias{new_output_component}
\title{Create scaffolding for a new Leptos Component}
\usage{
new_input_component(component_name, path = ".")

new_output_component(component_name, path = ".")
}
\arguments{
\item{component_name}{The name for the new component (e.g., "my_slider").
This will be converted to snake_case and CamelCase where needed.}

\item{path}{Path to the root of the R package. Default is the current directory.}
}
\description{
This function creates the necessary R, Rust, and TypeScript files
for a new input or output component based on a generic template.
}
