% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{get_pkg_name}
\alias{get_pkg_name}
\title{Get the name of the package from DESCRIPTION}
\usage{
get_pkg_name(path = ".")
}
\arguments{
\item{path}{Path to the package directory. Default is the current working directory.}
}
\value{
The name of the package as a string.
}
\description{
Get the name of the package from DESCRIPTION
}
