% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialize shiny.leptos structure in a package}
\usage{
init(path = ".")
}
\arguments{
\item{path}{Path to the root of the R package. Default is the current directory.}
}
\description{
Sets up the necessary directories and template files for using
shiny.leptos in an R package.
}
